resolved directory and postman issues
