+++ api/daily-journal.json
@@ -22,43 +22,11 @@
       "mood": "Ok"
     },
     {
+      "date": "10/30/2026",
+      "concepts": "JSON",
+      "entry": "json issues",
+      "mood": "Ok",
       "id": 4
-    },
-    {
-      "id": 5
-    },
-    {
-      "id": 6
-    },
-    {
-      "id": 7
-    },
-    {
-      "id": 8
-    },
-    {
-      "id": 9
-    },
-    {
-      "id": 10
-    },
-    {
-      "id": 11
-    },
-    {
-      "id": 12
-    },
-    {
-      "id": 13
-    },
-    {
-      "id": 14
-    },
-    {
-      "id": 15
-    },
-    {
-      "id": 16
     }
   ]
 }
