+++ src/index.html
@@ -113,7 +113,7 @@
 
     </footer>
 
-    <script type="module" src="./src/scripts/main.js"></script>
+    <script type="module" src="./scripts/main.js"></script>
 
 </body>
