+++ src/scripts/feed/journalEntry.js
