+++ src/scripts/main.js
@@ -1,11 +1,11 @@
 import { getEntries } from "./data/DataManager.js"
-import { EntryList } from "./feed/journalEntryList.js"
+import { EntryListComponent } from "./feed/journalEntryList.js"
 
 const showEntryList = () => {
     const postElement = document.querySelector(".entryLog");
     getEntries()
     .then((allEntries) => {
-        postElement.innerHTML = EntryList(allEntries);
+        postElement.innerHTML = EntryListComponent(allEntries);
     })
 }
